Remove application argument parsing from writeResultsToCSV() (#43)
It's easier to know what's going on if the function doesn't refer to
global variables to accepts the file path as a parameter.
