+++ src/index.ts
@@ -7,6 +7,9 @@ import {
     SimpleShippingMethod,
 } from '@lune-climate/lune'
 import cliProgress from 'cli-progress'
+import fs from 'fs'
+import minimist from 'minimist'
+import path from 'path'
 
 import { estimatePayload, EstimateResult, LegFromCSV } from './types.js'
 import {
@@ -222,9 +225,20 @@ async function main() {
     }
     progressBar.stop()
 
+    const argv = minimist(process.argv.slice(2))
+    const nameOfCSVFile = path.parse(process.argv[2]).name
+    const outputFilePath = `${argv.o || '.'}/${nameOfCSVFile}_${Date.now()}.csv`
+
+    if (argv.o && !fs.existsSync(argv.o)) {
+        fs.mkdirSync(argv.o, {
+            recursive: true,
+        })
+    }
+
     writeResultsToCSV({
         pathToShippingDataCSV: pathToCSVFile,
         results: estimates,
+        outputFilePath,
     })
 }
